requirements.txt
google-generativeai
requests
http.server